% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skill_forge_tools.R
\name{analyze_r_package}
\alias{analyze_r_package}
\title{Analyze R Package for Skill Creation}
\usage{
analyze_r_package(package)
}
\arguments{
\item{package}{Name of the package to analyze.}
}
\value{
A string summary of the package.
}
\description{
Introspects an installed R package to understand its capabilities, exported functions,
and documentation. This is used by the Skill Architect to "learn" a package.
}
